+++ render/linedrawer.py
@@ -20,7 +20,7 @@ class LineRenderer(RenderingInterface):
         self.real_genotype_size = self.genotype_size * self.num_lines
 
     def chunks(self, array):
-        array = torch.tensor(array)
+        array = torch.tensor(array, dtype=torch.float)
         return array.view(self.args.num_lines, self.genotype_size)
 
     def __str__(self):
