+++ render/pixeldrawer.py
@@ -180,7 +180,7 @@ class PixelRenderer(RenderingInterface):
         self.real_genotype_size = self.genotype_size * (args.num_lines * args.num_lines)
 
     def chunks(self, array):
-        array = torch.tensor(array)
+        array = torch.tensor(array, dtype=torch.float)
         return array.view(self.num_lines, self.num_lines, self.genotype_size)
 
     def __str__(self):
