+++ render/pylinhas.py
@@ -17,7 +17,7 @@ class PylinhasRenderer(RenderingInterface):
         self.real_genotype_size = self.genotype_size * args.num_lines
 
     def chunks(self, array):
-        array = torch.tensor(array)
+        array = torch.tensor(array, dtype=torch.float)
         return array.view(self.args.num_lines, self.genotype_size)
 
     def __str__(self):
